requirements.txt
-i https://pypi.org/simple
--extra-index-url https://download.pytorch.org/whl/cpu
amqp==5.2.0; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
billiard==4.2.0; python_version >= '3.7'
boto3==1.33.11
botocore==1.33.11; python_version >= '3.7'
celery==5.3.6
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
contourpy==1.2.0; python_version >= '3.9'
cycler==0.12.1; python_version >= '3.8'
django==4.2.5
django-cors-headers==4.3.1
django-jsonform==2.21.2
djangorestframework==3.14.0
filelock==3.13.1; python_version >= '3.8'
fonttools==4.46.0; python_version >= '3.8'
fsspec==2023.12.1; python_version >= '3.8'
gunicorn==21.2.0
idna==3.6; python_version >= '3.5'
jinja2==3.1.2; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
kiwisolver==1.4.5; python_version >= '3.7'
kombu==5.3.4; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
matplotlib==3.8.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.2.1; python_version >= '3.9'
numpy==1.26.2; python_version >= '3.9'
opencv-python==4.8.1.78; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.1.4; python_version >= '3.9'
pillow==10.1.0; python_version >= '3.8'
prompt-toolkit==3.0.41; python_full_version >= '3.7.0'
psutil==5.9.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
psycopg==3.1.14
psycopg-binary==3.1.14
py-cpuinfo==9.0.0
pyparsing==3.1.1; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.1
requests==2.31.0; python_version >= '3.7'
s3transfer==0.8.2; python_version >= '3.7'
scipy==1.11.4; python_version >= '3.9'
seaborn==0.13.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
sympy==1.12; python_version >= '3.8'
thop==0.1.1.post2209072238
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
tzdata==2023.3; python_version >= '2'
ultralytics==8.0.225
urllib3==2.0.7; python_version >= '3.10'
vine==5.1.0; python_version >= '3.6'
wcwidth==0.2.12
whitenoise==6.6.0
